\name{petBrainGlioma}
\docType{data}
\alias{petBrainGlioma}
\title{Standard Uptake Value (SUV) for Brain Glioma Grading}
\description{
  SUVs from FDG PET and ACE PET used in differentiating brain tumors.
 }
\usage{petBrainGlioma}
\format{
  A data-frame presenting the SUVs.
}
\source{
  Tatsuro Tsuchida, Hiroaki Takeuchi, Hidehiko Okazawa, Tetsuya
  Tsujikawa, and Yasuhisa Fujibayashi (2008)
  Grading of brain glioma with 1-$^{11}${C}-acetate {PET}: comparison with $^{18}${F}-{FDG PET}.
  \emph{Nuclear medicine and biology}
  \bold{35(2)} 171-176
}
\keyword{datasets}
